Solution to Check if a Number is Square of Another

+++ Check if a Number is Square of Another@@6788dd63490637cfc44e1ba4/code.c
@@ -5,8 +5,8 @@ char* welcome() {
 }
 
 int main() {
-    int a,b;
-    scanf("%d %d",&a,&b);
+    float a,b;
+    scanf("%f %f",&a,&b);
     if(a/b==b){
         printf("Yes");
     }
